requirements.txt
folium>=0.20.0
requests>=2.32.0
python-dotenv>=1.1.0
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=10.0.0
